fix xml-app fix
