--- xml-app/src/main/java/com/example/xml_app/MainActivity.kt
@@ -1,64 +1,0 @@
-/*
- *  Copyright (C) 2023 Rajesh Hadiya
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package com.example.xml_app
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.fragment.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
-import com.example.xml_app.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    lateinit var navController: NavController
-    private lateinit var appbarConfiguration: AppBarConfiguration
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
-
-        val navHostFragment =
-            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
-        navController = navHostFragment.findNavController()
-
-        /* this is used to mention that these two fragments are the topfragments in the appbarnavigation , and backbutton
-        shoudnt be shown on navigating into these bottomNavbar fragments.*/
-        appbarConfiguration =
-            AppBarConfiguration(setOf(R.id.profileListFragment), binding.navDrawer)
-
-        setSupportActionBar(binding.toolbar)
-        // applying the appbarconfiguration to avoid the backbutton in the toolbar/actionbar
-        setupActionBarWithNavController(navController, appbarConfiguration)
-
-        binding.navigationView.setupWithNavController(navController)
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return navController.navigateUp(appBarConfiguration = appbarConfiguration) || super.onSupportNavigateUp()
-    }
-}
