--- xml-app/src/main/java/com/example/xml_app/ProfileApplication.kt
@@ -1,7 +1,0 @@
-package com.example.xml_app
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class ProfileApplication : Application()
