--- xml-app/src/main/java/com/example/xml_app/util/UiState.kt
@@ -1,11 +1,0 @@
-package com.example.xml_app.util
-
-sealed class UiState<out T> {
-    object Empty : UiState<Nothing>()
-
-    object Loading : UiState<Nothing>()
-
-    data class Success<out T>(val data: T) : UiState<T>()
-
-    data class Error(val data: String?) : UiState<Nothing>()
-}
