+++ xml-app/src/main/java/com/hadiyarajesh/xml_app/ProfileApplication.kt
@@ -1,0 +1,7 @@
+package com.hadiyarajesh.xml_app
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class ProfileApplication : Application()
