+++ xml-app/src/main/java/com/hadiyarajesh/xml_app/database/FlowerDatabase.kt
@@ -1,0 +1,31 @@
+/*
+ *  Copyright (C) 2023 Rajesh Hadiya
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.hadiyarajesh.xml_app.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+import com.hadiyarajesh.xml_app.database.entity.Image
+import com.hadiyarajesh.xml_app.database.dao.ImageDao
+
+@Database(
+    version = 1,
+    exportSchema = true,
+    entities = [Image::class]
+)
+abstract class FlowerDatabase : RoomDatabase() {
+    abstract val imageDao: ImageDao
+}
