+++ xml-app/src/main/java/com/hadiyarajesh/xml_app/database/dao/ImageDao.kt
@@ -1,0 +1,39 @@
+/*
+ *  Copyright (C) 2023 Rajesh Hadiya
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.hadiyarajesh.xml_app.database.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.hadiyarajesh.xml_app.database.entity.Image
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface ImageDao {
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun insertAllImages(images: List<Image>)
+
+    @Query("SELECT * FROM Image ORDER BY id")
+    fun getAllImages(): Flow<List<Image>>
+
+    @Query("SELECT * FROM Image WHERE id=:imageId ORDER BY id")
+    fun getImageById(imageId: Long): Flow<Image>
+
+    @Query("DELETE FROM Image")
+    suspend fun deleteAllImages()
+}
