+++ xml-app/src/main/java/com/hadiyarajesh/xml_app/database/entity/Image.kt
@@ -1,0 +1,19 @@
+package com.hadiyarajesh.xml_app.database.entity
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@Entity
+@JsonClass(generateAdapter = true)
+data class Image(
+    @PrimaryKey
+    val id: String,
+    val author: String,
+    val width: Int,
+    val height: Int,
+    val url: String,
+    @Json(name = "download_url")
+    val downloadUrl: String
+)
