+++ xml-app/src/main/java/com/hadiyarajesh/xml_app/di/DatabaseModule.kt
@@ -1,0 +1,47 @@
+/*
+ *  Copyright (C) 2023 Rajesh Hadiya
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package com.hadiyarajesh.xml_app.di
+
+import android.content.Context
+import androidx.room.Room
+import com.hadiyarajesh.xml_app.R
+import com.hadiyarajesh.xml_app.database.FlowerDatabase
+import com.hadiyarajesh.xml_app.database.dao.ImageDao
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+class DatabaseModule {
+    @Provides
+    @Singleton
+    fun provideFlowerDatabase(@ApplicationContext context: Context): FlowerDatabase {
+        return Room.databaseBuilder(
+            context.applicationContext,
+            FlowerDatabase::class.java,
+            "${context.getString(R.string.app_name)}.db"
+        ).build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideImageDao(db: FlowerDatabase): ImageDao = db.imageDao
+}
