+++ xml-app/src/main/res/layout/activity_main.xml
@@ -1,7 +1,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity">
+    tools:context="com.hadiyarajesh.xml_app.MainActivity">
 
     <androidx.drawerlayout.widget.DrawerLayout
         android:id="@+id/nav_drawer"
