+++ xml-app/src/main/res/layout/fragment_profile_detail.xml
@@ -13,7 +13,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:orientation="vertical"
-            tools:context=".fragment.detail.ProfileDetailFragment"
+            tools:context="com.hadiyarajesh.xml_app.fragment.detail.ProfileDetailFragment"
             android:visibility="visible">
 
             <com.google.android.material.card.MaterialCardView
