+++ xml-app/src/main/res/layout/fragment_profile_list.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".fragment.home.ProfileListFragment">
+        tools:context="com.hadiyarajesh.xml_app.fragment.home.ProfileListFragment">
 
         <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
             android:id="@+id/swipeRefreshLayout"
