+++ xml-app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/profileListFragment"
-        android:name="com.example.xml_app.fragment.home.ProfileListFragment"
+        android:name="com.hadiyarajesh.xml_app.fragment.home.ProfileListFragment"
         android:label="Profile"
         tools:layout="@layout/fragment_profile_list">
         <action
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/profileDetailFragment"
-        android:name="com.example.xml_app.fragment.detail.ProfileDetailFragment"
+        android:name="com.hadiyarajesh.xml_app.fragment.detail.ProfileDetailFragment"
         android:label="@string/details"
         tools:layout="@layout/fragment_profile_detail" >
         <argument
